# backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Scambio di figurine",
    "description": "Applicazione web che permette lo scambio, acquisto, e apertura di pacchetti di figurine Marvel",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "Gestione Utenti"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addUserSchema"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Gestione Utenti"
        ],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "tags": [
          "Gestione Utenti"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateUserSchema"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Gestione Utenti"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "get": {
        "tags": [
          "Gestione Utenti"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/credits/{id}/{qty}": {
      "put": {
        "tags": [
          "Gestione Crediti Utente"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "qty",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/credits/{id}": {
      "get": {
        "tags": [
          "Gestione Crediti Utente"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/figurine/{id}": {
      "get": {
        "tags": [
          "Gestione Figurine"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Gestione Figurine"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/putFigurineSchema"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Gestione Figurine"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/postFigurineSchema"
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginSchema"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "addUserSchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Gabriele"
          },
          "surname": {
            "type": "string",
            "example": "Cucchi"
          },
          "age": {
            "type": "number",
            "example": 18
          },
          "hero": {
            "type": "string",
            "example": "Batman"
          },
          "email": {
            "type": "string",
            "example": "gabriele.cucchi@studenti.unimi.it"
          },
          "psw": {
            "type": "string",
            "example": "password"
          },
          "credits": {
            "type": "number",
            "example": 0
          },
          "figurine": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "count": {
                  "type": "number",
                  "example": 2
                }
              }
            }
          }
        },
        "required": [
          "name",
          "surname",
          "email",
          "psw",
          "credits"
        ]
      },
      "updateUserSchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Gabriele"
          },
          "surname": {
            "type": "string",
            "example": "Cucchi"
          },
          "age": {
            "type": "number",
            "example": 18
          },
          "hero": {
            "type": "string",
            "example": "Superman"
          },
          "psw": {
            "type": "string",
            "example": "password"
          },
          "credits": {
            "type": "number",
            "example": 0
          }
        }
      },
      "putFigurineSchema": {
        "type": "object",
        "properties": {
          "figurine": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "count": {
                  "type": "number",
                  "example": 2
                }
              }
            }
          }
        },
        "required": [
          "figurine"
        ]
      },
      "postFigurineSchema": {
        "type": "object",
        "properties": {
          "figurine": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 2
                },
                "count": {
                  "type": "number",
                  "example": 3
                }
              }
            }
          }
        },
        "required": [
          "figurine"
        ]
      },
      "loginSchema": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "gabriele.cucchi@studenti.unimi.it"
          },
          "psw": {
            "type": "string",
            "example": "password"
          }
        },
        "required": [
          "email",
          "psw"
        ]
      }
    }
  }
}